feat(network): complete network layer & REST repositories integration for Client, Admin, Location, and Rating
- Implemented ClientOrderRepository with REST endpoints for order creation, list, detail, and cancellation.
- Implemented LocationApiService & LocationRepository with autocomplete & route estimation endpoints.
- Implemented AdminApiService & AdminRepository with user, driver, alert, and order management endpoints.
- Integrated RatingApiService into RetrofitClient and updated RatingRepository/RatingViewModel.
- Migrated OrderViewModel, CustomerViewModel, and AdminViewModel from Room local storage to REST Repositories.
- Updated UI screens (OrderTrackingScreen, CustomerOrderDetailScreen, CustomerOrderListScreen) to consume domain models.
- Deleted deprecated ApiClient and root RatingApiService files.

+++ Code/app/src/main/java/com/mob10/deliveryapp/ui/rating/RatingDialog.kt
@@ -34,9 +34,9 @@ import androidx.compose.ui.unit.dp
  */
 @Composable
 fun RatingDialog(
-    deliveryRequestId: Int,
-    clientId: Int,
-    driverId: Int,
+    deliveryRequestId: Long,
+    clientId: Long,
+    driverId: Long,
     onDismiss: () -> Unit,
     onSubmit: (stars: Int, comment: String) -> Unit,
     isSubmitting: Boolean = false,
